Revert "Move common ovsp4rt files to 'common' subdirectory (#467)"
This reverts commit b74d83d231529840fb2659f4cab66ea9b50cf353.
