+++ ovs-p4rt/sidecar/ovsp4rt.cc
@@ -6,9 +6,9 @@
 #include <string>
 
 #include "absl/flags/flag.h"
-#include "common/ovsp4rt_credentials.h"
-#include "common/ovsp4rt_session.h"
 #include "openvswitch/ovs-p4rt.h"
+#include "ovsp4rt_credentials.h"
+#include "ovsp4rt_session.h"
 
 #if defined(DPDK_TARGET)
 #include "dpdk/p4_name_mapping.h"
